# Performance tuning users.php (rewrite a query + add index)