--- a/src/include/data_access.php
+++ b/src/include/data_access.php
@@ -590,10 +590,13 @@
       }
 
       $ret = array();
-      $sql = "SELECT * FROM `". DB_MAP_TABLE ."` a ".
-             "INNER JOIN `". DB_MAP_TABLE ."` b ".
-             "ON a.ID=b.ID ".
-             "WHERE a.`$field`=(SELECT MAX(`$field`) FROM `". DB_MAP_TABLE ."` WHERE UserID=b.UserID AND (ProtectedUntil IS NULL OR ProtectedUntil<='$now' OR UserID='$requestingUserID'))";
+      $sql = "SELECT maps.* FROM `". DB_MAP_TABLE ."` maps " .
+             "INNER JOIN (".
+             "    SELECT UserID, MAX(`$field`) AS value" .
+             "    FROM `". DB_MAP_TABLE ."`" .
+             "    WHERE (ProtectedUntil IS NULL OR ProtectedUntil<='$now' OR UserID='$requestingUserID')" .
+             "    GROUP BY UserID" .
+             ") VI ON maps.UserID = VI.UserID AND maps.`$field` = VI.value";
       $rs = self::Query($sql);
       while($r = mysqli_fetch_assoc($rs))
       {