- We have about 5500 maps in the database and the frontpage (users.php) took about 10 seconds to load.
- GetLastMapsPerUser took almost 7sec and GetLastComments took 1,7sec
- I rewrote the SQL for GetLastMapsPerUser to use an inline view with a GROUP BY and equivalent timing and access control as the original query.
- I compared the output and got the exact same result as before. 
- I also added an index on the doma_comments table (for the next doma version).
- You can check out the performance at: https://kartarkiv.nydalen.idrett.no/ (0,5sec response time)

Here is a concrete version of the SQL query:
SELECT maps.* FROM doma_maps maps
INNER JOIN (
	SELECT UserID, MAX(Date) AS value
    FROM doma_maps
    WHERE (ProtectedUntil IS NULL OR ProtectedUntil<='2020-01-21' OR UserID='0')
    GROUP BY UserID
) VI ON maps.UserID = VI.UserID AND maps.Date = VI.value